Classes Interval and RangeOfIntervals
Added two classes:

Interval which is basically two ordered floats with comparison oparator
and merging method (beware : equality means intersection)

RangeOfIntervals: length of the range, boolean if sorted, and list of
intervals. When added an interval is merged if it intersect a interval
already in the range. Use method append(interval) to add an Interval.