Added Discretize method for GraphicalIntervals
GraphicalIntervals Discretize method override SetOfIntervals by calling
the latter and plotting its return value. The format of this return
value has been changed to represent the characteristic function
described in README and help. It is a duple of lists, first list with x
values, second list with boolean values of the characteristic function.