Implemented discretization for GraphicIntervals
The discretization is showed on the plot.
Added a attribute containing the discretization parameters (allows
multiple calls). Method `changeDiscretizationParameters` with tuple or
list argument to change the discretization parameters.
Now the GraphicalIntervals class can recieve an argument for
displaybutton to specify if a discretization button is to be plotted
(default True).