Add import of MeasuredValues for Class Case
Method case.findtimes fetch the tb (time beginning) and te (time end).
Optional arguments are : PathToFile (where to look for the measurement
file (default is for my computer) and AA which is an instance of the
class measuredValues. Please prefer giving AA to fetching. If many cases
are to be created, the average time to call AA is 0.7 s.
Added method getVariable(ID,mic,variable) to get a variable from
MeasuredValues (for example 'Tb' or 'Te').