Some fixes test
There might be a lagging problem with the app. When one lets it runs for
a few minutes (even without playing audio or selecting ranges) the app
is lagging, and all the system is very slow. The app disappears and then
reappears but non responsive. Then it cannot be quickly closed. I tryed
a few fixes (displayed in this commit but it does work)