Fixed the bug
Know the timer is stopped when the audio is not playing. This avoid the
freezing when Qt takes is a bit late in updating all the files. The
spans are still updated even if the audio is not playing. The only
downside is that the span is not appearing before release of mouse
button (when audio off). This could be corrected, but I don't think it
is urgent.