Multiple Paths search
Implemented a multiple path search for Micsignal initialisation. This
way, we can keep the .mat objects in the folder they are currently in.
The paths can either be given directly in the algorithm_evaluation.py or
through a GUI  (callGUI has to be set to True in
algorithm_evaluation.py)